requirements.txt
fastapi
uvicorn
pillow
exifread